Show already voted status frontend
If you already voted, you should be directed to look at the results
instead.

Should we show the link to the results at all times?

merged in with this request: a bugfix for social sharing, it would not
work if you previously selected something in a page because it didn't
clear your selection.